Resolve #2002: dedupe i18n catalog helpers
