+++ packages/i18n/src/icu.ts
@@ -1,9 +1,10 @@
 import { FormatError, IntlMessageFormat } from 'intl-messageformat';
 import type { Formats } from 'intl-messageformat';
 
+import { resolveCatalogMessage } from './catalog.js';
 import { I18nError } from './errors.js';
 import { I18nService, createI18n } from './service.js';
-import type { I18nInterpolationValues, I18nLocale, I18nMessageTree, I18nModuleOptions, I18nTranslateOptions } from './types.js';
+import type { I18nInterpolationValues, I18nLocale, I18nModuleOptions, I18nTranslateOptions } from './types.js';
 
 /**
  * Primitive values accepted by the ICU MessageFormat subpath.
@@ -55,39 +56,12 @@ function toMessageFormatValues(values: I18nIcuValues | undefined): Record<string
   return Object.fromEntries(Object.entries(values));
 }
 
-function hasOwn(value: unknown, key: string): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && Object.hasOwn(value, key);
-}
-
-function resolveMessage(tree: I18nMessageTree | undefined, key: string): string | undefined {
-  if (tree === undefined) {
-    return undefined;
-  }
-
-  if (hasOwn(tree, key)) {
-    const direct = tree[key];
-    return typeof direct === 'string' ? direct : undefined;
-  }
-
-  let current: unknown = tree;
-
-  for (const part of key.split('.')) {
-    if (!hasOwn(current, part)) {
-      return undefined;
-    }
-
-    current = current[part];
-  }
-
-  return typeof current === 'string' ? current : undefined;
-}
-
 function resolveMessageLocale(service: I18nService, key: string, options: I18nIcuTranslateOptions): I18nLocale {
   const resolvedKey = options.namespace === undefined ? key : `${options.namespace}.${key}`;
   const snapshot = service.snapshotOptions();
 
   for (const locale of service.resolveLocales(options.locale)) {
-    if (resolveMessage(snapshot.catalogs?.[locale], resolvedKey) !== undefined) {
+    if (resolveCatalogMessage(snapshot.catalogs?.[locale], resolvedKey) !== undefined) {
       return locale;
     }
   }
