+++ packages/i18n/src/loaders/shared.ts
@@ -1,4 +1,5 @@
 import { I18nError } from '../errors.js';
+import { isPlainObject, snapshotMessageTree } from '../catalog.js';
 import type { I18nLocale, I18nMessageTree, I18nTranslationKey } from '../types.js';
 
 const SAFE_LOCALE_PATTERN = /^[A-Za-z0-9](?:[A-Za-z0-9_-]*[A-Za-z0-9])?(?:-[A-Za-z0-9](?:[A-Za-z0-9_-]*[A-Za-z0-9])?)*$/;
@@ -27,20 +28,7 @@ export interface I18nLoader {
   load(locale: I18nLocale, namespace: I18nTranslationKey, options?: I18nLoaderLoadOptions): Promise<I18nMessageTree>;
 }
 
-/**
- * Checks whether a value is a plain object suitable for loader option or catalog validation.
- *
- * @param value Candidate value to inspect.
- * @returns `true` when the value is a plain object or null-prototype object.
- */
-export function isPlainObject(value: unknown): value is Record<string, unknown> {
-  if (typeof value !== 'object' || value === null) {
-    return false;
-  }
-
-  const prototype = Object.getPrototypeOf(value);
-  return prototype === Object.prototype || prototype === null;
-}
+export { isPlainObject } from '../catalog.js';
 
 /**
  * Validates a loader locale before a backend is called.
@@ -87,29 +75,5 @@ export function validateLoaderNamespace(namespace: unknown, label: string): asse
  * @returns A frozen message tree detached from caller-owned data.
  */
 export function snapshotLoaderMessageTree(value: unknown, path: string): I18nMessageTree {
-  if (!isPlainObject(value)) {
-    throw new I18nError(`${path} must be a plain object message tree.`, 'I18N_INVALID_CATALOG');
-  }
-
-  const snapshot: Record<string, string | I18nMessageTree> = {};
-
-  for (const [key, entry] of Object.entries(value)) {
-    if (key.trim() === '') {
-      throw new I18nError(`${path} contains an empty message key segment.`, 'I18N_INVALID_CATALOG');
-    }
-
-    if (typeof entry === 'string') {
-      snapshot[key] = entry;
-      continue;
-    }
-
-    if (isPlainObject(entry)) {
-      snapshot[key] = snapshotLoaderMessageTree(entry, `${path}.${key}`);
-      continue;
-    }
-
-    throw new I18nError(`${path}.${key} must be a string or nested message tree.`, 'I18N_INVALID_CATALOG');
-  }
-
-  return Object.freeze(snapshot);
+  return snapshotMessageTree(value, path);
 }
